requirements.txt
google-generativeai
streamlit
python-dotenv
chromadb
sentence-transformers